what

+++ libft_utils.c
@@ -6,15 +6,15 @@
 /*   By: stmuller <stmuller@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/10 16:50:17 by stmuller          #+#    #+#             */
-/*   Updated: 2026/01/19 23:44:03 by stmuller         ###   ########.fr       */
+/*   Updated: 2026/01/20 01:39:45 by stmuller         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minitalk.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
